Rework node handling
- Separate NodeProvider to type-specific node
  providers, which now also handle decompilation
- Introduce ILSpyXApplication as future holder of
  for backend application state
