+++ backend/src/ILSpy.Backend/Decompiler/IDecompilerBackend.cs
@@ -12,8 +12,6 @@ public interface IDecompilerBackend
 {
     AssemblyData? AddAssembly(string? path);
     bool RemoveAssembly(string? path);
-    IEnumerable<string> ListAssemblyReferences(string? assemblyPath);
-    IEnumerable<string> ListNamespaces(string? assemblyPath);
     IEnumerable<MemberData> ListTypes(string? assemblyPath, string? @namespace);
     IDictionary<string, string> GetCode(string? assemblyPath, EntityHandle handle);
     IEnumerable<MemberData> GetMembers(string? assemblyPath, TypeDefinitionHandle handle);
