+++ backend/src/ILSpy.Backend/Handlers/AddAssemblyHandler.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See the LICENSE file in the project root for more information.
 
+using ILSpy.Backend.Application;
 using ILSpy.Backend.Decompiler;
 using ILSpy.Backend.Protocol;
 using OmniSharp.Extensions.JsonRpc;
@@ -12,12 +13,12 @@ namespace ILSpy.Backend.Handlers
     [Serial, Method("ilspy/addAssembly", Direction.ClientToServer)]
     public class AddAssemblyHandler : IJsonRpcRequestHandler<AddAssemblyRequest, AddAssemblyResponse>
     {
-        private readonly IDecompilerBackend decompilerBackend;
+        private readonly ILSpyXApplication application;
         private readonly SearchBackend searchBackend;
 
-        public AddAssemblyHandler(IDecompilerBackend decompilerBackend, SearchBackend searchBackend)
+        public AddAssemblyHandler(ILSpyXApplication application, SearchBackend searchBackend)
         {
-            this.decompilerBackend = decompilerBackend;
+            this.application = application;
             this.searchBackend = searchBackend;
         }
 
@@ -28,7 +29,7 @@ namespace ILSpy.Backend.Handlers
                 await searchBackend.AddAssembly(request.AssemblyPath);
             }
 
-            var result = request.AssemblyPath != null ? decompilerBackend.AddAssembly(request.AssemblyPath) : null;
+            var result = request.AssemblyPath != null ? application.DecompilerBackend.AddAssembly(request.AssemblyPath) : null;
             return new AddAssemblyResponse(Added: result != null, AssemblyData: result);
         }
     }
